backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
pydantic[email]==2.10.4
pydantic-settings==2.7.0
python-docx==1.1.2
pdf2docx
PyMuPDF
numpy
Pillow
paddleocr
paddlepaddle
